pyproject.toml
[project]
name = "sample-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.132.0",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "pwdlib[argon2]>=0.3.0",
]
